refactor: 后端增加按钮权限的校验；权限校验仅在controller层校验

+++ anno-admin/src/main/java/site/sorghum/anno/anno/controller/BaseDbController.java
@@ -15,13 +15,15 @@ import site.sorghum.anno._common.util.CryptoUtil;
 import site.sorghum.anno._common.util.JSONUtil;
 import site.sorghum.anno._metadata.AnEntity;
 import site.sorghum.anno._metadata.AnField;
+import site.sorghum.anno._metadata.PermissionContext;
 import site.sorghum.anno._metadata.MetadataManager;
 import site.sorghum.anno.anno.entity.common.AnnoTreeDTO;
+import site.sorghum.anno.anno.interfaces.CheckPermissionFunction;
+import site.sorghum.anno.anno.proxy.PermissionProxy;
 import site.sorghum.anno.anno.util.AnnoTableParamCache;
 import site.sorghum.anno.anno.util.AnnoUtil;
 import site.sorghum.anno.anno.util.Utils;
 import site.sorghum.anno.db.param.DbCondition;
-import site.sorghum.anno.db.param.OrderByParam;
 import site.sorghum.anno.db.param.PageParam;
 import site.sorghum.anno.db.param.TableParam;
 import site.sorghum.anno.db.service.DbService;
@@ -45,6 +47,10 @@ public class BaseDbController {
 
     @Inject
     MetadataManager metadataManager;
+    @Inject
+    PermissionContext permissionContext;
+    @Inject
+    PermissionProxy permissionProxy;
 
     /**
      * 分页查询
@@ -59,7 +65,11 @@ public class BaseDbController {
                                          boolean ignoreM2m,
                                          boolean reverseM2m,
                                          Map<String, Object> param) {
+
         AnEntity entity = metadataManager.getEntity(clazz);
+
+        permissionProxy.checkPermission(entity, PermissionProxy.VIEW);
+
         param = emptyStringIgnore(param);
         String m2mSql = Utils.m2mSql(param);
         String andSql = null;
@@ -83,18 +93,23 @@ public class BaseDbController {
     }
 
     public <T> AnnoResult<T> save(String clazz, Map<String, Object> param) {
+        permissionProxy.checkPermission(metadataManager.getEntity(clazz), PermissionProxy.ADD);
+
         TableParam<T> tableParam = (TableParam<T>) AnnoTableParamCache.get(clazz);
+
         T t = JSONUtil.toBean(emptyStringIgnore(param), tableParam.getClazz());
         dbService.insert(t);
         return AnnoResult.succeed();
     }
 
     public <T> AnnoResult<T> queryById(String clazz, String pkValue, String _cat) {
+        AnEntity anEntity = metadataManager.getEntity(clazz);
+        permissionProxy.checkPermission(anEntity, PermissionProxy.VIEW);
+
         String id = pkValue;
         if (id == null) {
             id = _cat;
         }
-        AnEntity anEntity = metadataManager.getEntity(clazz);
         AnField pkField = anEntity.getPkField();
         if (pkField == null) {
             return AnnoResult.failure("未找到主键");
@@ -112,6 +127,8 @@ public class BaseDbController {
      */
     public AnnoResult<String> removeById(String clazz, String id) {
         AnEntity anEntity = metadataManager.getEntity(clazz);
+        permissionProxy.checkPermission(anEntity, PermissionProxy.DELETE);
+
         AnField pkField = anEntity.getPkField();
         dbService.delete(anEntity.getClazz(), CollUtil.newArrayList(DbCondition.builder().field(pkField.getTableFieldName()).value(id).build()));
         return AnnoResult.succeed("删除成功");
@@ -122,8 +139,9 @@ public class BaseDbController {
      */
     public <T> AnnoResult<T> updateById(String clazz, Map<String, Object> param) {
         AnEntity anEntity = metadataManager.getEntity(clazz);
+        permissionProxy.checkPermission(anEntity, PermissionProxy.UPDATE);
+
         AnField pkField = anEntity.getPkField();
-        TableParam<T> tableParam = metadataManager.getTableParam(clazz);
         T bean = (T) JSONUtil.toBean(emptyStringIgnore(param), anEntity.getClazz());
         dbService.update(CollUtil.newArrayList(DbCondition.builder().field(pkField.getTableFieldName()).value(param.get(pkField.getFieldName())).build()), bean);
         return AnnoResult.succeed();
@@ -132,20 +150,23 @@ public class BaseDbController {
     public <T> AnnoResult<T> saveOrUpdate(String clazz, Map<String, Object> param) {
         AnEntity anEntity = metadataManager.getEntity(clazz);
         T data = (T) JSONUtil.toBean(emptyStringIgnore(param), anEntity.getClazz());
-        TableParam<T> tableParam = (TableParam<T>) metadataManager.getTableParam(clazz);
         AnField pkField = anEntity.getPkField();
         if (pkField == null) {
             return AnnoResult.failure("未找到主键");
         }
         if (ReflectUtil.getFieldValue(data, pkField.getFieldName()) == null) {
+            permissionProxy.checkPermission(anEntity, PermissionProxy.ADD);
             dbService.insert(data);
         } else {
+            permissionProxy.checkPermission(anEntity, PermissionProxy.UPDATE);
             dbService.update(CollUtil.newArrayList(DbCondition.builder().field(pkField.getTableFieldName()).value(param.get(pkField.getFieldName())).build()), data);
         }
         return AnnoResult.succeed(data);
     }
 
     public <T> AnnoResult<T> removeRelation(String clazz, Map<String, String> param) throws SQLException {
+        permissionProxy.checkPermission(metadataManager.getEntity(clazz), PermissionProxy.DELETE);
+
         String mediumOtherField = param.get("mediumOtherField");
         String otherValue = param.get("joinValue");
         String thisValue = param.get(param.get("joinThisClazzField"));
@@ -162,6 +183,15 @@ public class BaseDbController {
                                                                boolean ignoreM2m,
                                                                boolean reverseM2m,
                                                                Map<String, String> param) {
+        List<T> list = queryTreeList(clazz, ignoreM2m, reverseM2m, param);
+        List<AnnoTreeDTO<String>> annoTreeDTOs = Utils.toTrees(list);
+        annoTreeDTOs.add(0, AnnoTreeDTO.<String>builder().id("0").label("无选择").value("").build());
+        return AnnoResult.succeed(annoTreeDTOs);
+    }
+
+    private <T> List<T> queryTreeList(String clazz, boolean ignoreM2m, boolean reverseM2m, Map<String, String> param) {
+        permissionProxy.checkPermission(metadataManager.getEntity(clazz), PermissionProxy.VIEW);
+
         TableParam<T> tableParam = (TableParam<T>) metadataManager.getTableParam(clazz);
         String m2mSql = Utils.m2mSql(param);
         String andSql = null;
@@ -178,39 +208,24 @@ public class BaseDbController {
             dbConditions.add(DbCondition.builder().type(DbCondition.QueryType.CUSTOM).field(andSql).build());
         }
         List<T> list = dbService.list(tableParam.getClazz(), dbConditions);
-        List<AnnoTreeDTO<String>> annoTreeDTOs = Utils.toTrees(list);
-        annoTreeDTOs.add(0, AnnoTreeDTO.<String>builder().id("0").label("无选择").value("").build());
-        return AnnoResult.succeed(annoTreeDTOs);
+        return list;
     }
 
     public <T> AnnoResult<Map<?, ?>> annoTreeSelectData(String clazz,
                                                         boolean ignoreM2m,
                                                         boolean reverseM2m,
                                                         Map<String, String> param) {
-        TableParam<T> tableParam = (TableParam<T>) metadataManager.getTableParam(clazz);
-        String m2mSql = Utils.m2mSql(param);
-        String andSql = null;
-        String inPrefix = " in (";
-        if (reverseM2m) {
-            inPrefix = " not in (";
-        }
-        if (StrUtil.isNotEmpty(m2mSql) && !ignoreM2m) {
-            String joinThisClazzField = param.get("joinThisClazzField");
-            andSql = joinThisClazzField + inPrefix + m2mSql + ")";
-        }
-        List<DbCondition> dbConditions = AnnoUtil.simpleEntity2conditions(param, tableParam.getClazz());
-        if (andSql != null) {
-            dbConditions.add(DbCondition.builder().type(DbCondition.QueryType.CUSTOM).field(andSql).build());
-        }
-        List<T> list = dbService.list(tableParam.getClazz(), dbConditions);
+        List<Object> list = queryTreeList(clazz, ignoreM2m, reverseM2m, param);
         if (list == null || list.isEmpty()) {
             return AnnoResult.succeed(Collections.emptyMap());
         }
+        TableParam<T> tableParam = (TableParam<T>) metadataManager.getTableParam(clazz);
         List<Object> data = list.stream().map(item -> ReflectUtil.getFieldValue(item, AnnoUtil.getPkField(tableParam.getClazz()))).collect(Collectors.toList());
         return AnnoResult.succeed(MapUtil.of("m2mTree", data));
     }
 
     public <T> AnnoResult<String> addM2m(String clazz, Map param, boolean clearAll) {
+        permissionProxy.checkPermission(metadataManager.getEntity(clazz), PermissionProxy.ADD);
         // 中间表
         String mediumTableClass = param.get("mediumTableClass").toString();
         Class<?> aClass = metadataManager.getEntity(mediumTableClass).getClazz();
@@ -249,13 +264,16 @@ public class BaseDbController {
         map.put("clazz", CryptoUtil.decrypt(map.get("clazz")));
         map.put("method", CryptoUtil.decrypt(map.get("method")));
         map.put("expireTime", CryptoUtil.decrypt(map.get("expireTime")));
+        // 校验权限
+        String permissionCode = permissionContext.getPermissionCode(map.get("clazz"), map.get("method"));
+        CheckPermissionFunction.permissionCheckFunction.accept(permissionCode);
+
         // 判断是否过期
         if (Long.parseLong(map.get("expireTime")) < System.currentTimeMillis()) {
             return AnnoResult.failure("页面已过期，请刷新页面重试。");
         }
-        Object bean = AnnoBeanUtils.getBean(
-            Class.forName(map.get("clazz"))
-        );
+
+        Object bean = AnnoBeanUtils.getBean(Class.forName(map.get("clazz")));
         ReflectUtil.invoke(bean, map.get("method"), map);
         return AnnoResult.succeed("执行成功");
     }
